scraping/exercises/50shades/requirements/base.txt
Django==1.6
psycopg2==2.5.1

# http://django-configurations.readthedocs.org/en/latest/cookbook/#envdir
# envdir==0.5

django-configurations==0.6
django-cache-url==0.6.0
dj-database-url==0.2.2

South==0.8.2